syllabus_question_generator/requirements.txt
Flask==3.0.3
python-docx==1.1.0
PyPDF2==3.0.1
reportlab==4.1.0
